requirements.txt
discord.py>=2.0.0
psycopg2-binary>=2.9.3
python-dotenv>=0.20.0
pytz>=2022.1
